[Test] Move concurrent activate update test to run with single node (#90561)
To ensure all version conflict errors are succussfully handled
internally, the test needs to run with a single node cluster. This is
because reading the same document from the primary shard and the replica
shard can get different result. In this case, the primary shard has the
updated document and reports version conflict, but the subsequent read
from the replica shard still sees the old document. This leads to
equality comparison to fail and in turn causes the verison conflict
error to be rethrown.

Resolves: #89776
